tv 바뀐 response 타입 적용

+++ tv/src/main/java/com/danbam/tv/ui/home/HomeViewModel.kt
@@ -32,7 +32,7 @@ class HomeViewModel @Inject constructor(
         }
     }
 
-    fun saveCurrentIndex(index: Int) = intent {
+    fun saveCurrentIndex(index: Long) = intent {
         reduce { state.copy(currentMovieIndex = index) }
     }
