+++ tv/src/main/java/com/danbam/tv/ui/movie/movie/MovieContract.kt
@@ -5,6 +5,6 @@ import com.danbam.domain.entity.MovieEntity
 import kotlinx.coroutines.flow.Flow
 
 data class MovieState(
-    val currentMovieIndex: Int = 0,
+    val currentMovieIndex: Long = 0,
     val movieAllPager: Flow<PagingData<MovieEntity>>? = null,
 )
